PROPOSAL: Formatting Fixes

+++ Engine/Core/Scene/Scene.h
@@ -1,4 +1,8 @@
 #pragma once
+#include <iostream>
+#include <vector>
+#include <ostream>
+
 #include "Engine/Audio/Audio.h"
 #include "Engine/Core/AssetManager.h"
 #include "Engine/Core/GameObject.h"
@@ -10,20 +14,12 @@
 #include "Engine/Game/Door.h"
 #include "Engine/Game/Gun.h"
 #include "Engine/Core/Animation.h"
-#include <iostream>
-#include <vector>
-#include <ostream>
 #include "Engine/Game/Door.h"
 #include "Engine/Game/Gun.h"
 #include "Engine/Game/Crate.h"
 #include "Engine/Game/Window.h"
-#include "Engine/Core/Lights/light.h"
-
+#include "Engine/Core/Lights/Light.h"
 #include "Engine/Physics/BulletPhysics.h"
-
-
-
-
  
 class Scene
 {
@@ -38,18 +34,14 @@ public:
 	void AddGunPickUp(GunPickUp gunpickup);
 	void AddGunPickUp(std::string gunName,std::string gunObject,glm::vec3 Position);
 
-	int GetGunPickUpSize();
+	size_t GetGunPickUpSize();
 	Crate* GetCrate(std::string name);
-
 private:
 	SkyBox sky;
-	//objects
+	// Objects
 	std::vector<Door> doors;
 	std::vector<Crate> crates;
 	std::vector<GunPickUp> gunPickUps;
 	std::vector<Window> windows;
 	std::vector<Light> lights;
-
-	
 };
-
