+++ Engine/Core/Scene/SceneManager.cpp
@@ -1,18 +1,19 @@
 #include "SceneManager.h"
 
-
-namespace SceneManager {
-
+namespace SceneManager
+{
 	std::vector<Scene> scenes;
 	int currentScene;
 
 	void SceneManager::Init() {
 		currentScene = 0;
 	}
-	int SceneManager::CreateScene(Scene scene) {
+
+	size_t SceneManager::CreateScene(Scene scene) {
 		scenes.push_back(scene);
 		return scenes.size() - 1;
 	}
+	
 	void SceneManager::LoadScene(int index) {
 		if (index < scenes.size())
 			currentScene = index;
@@ -23,6 +24,7 @@ namespace SceneManager {
 	void SceneManager::Update(float dt) {
 		scenes[currentScene].Update(dt);
 	}
+	
 	void SceneManager::Render() {
 		scenes[currentScene].RenderObjects();
 	}
